+++ src/ts/wysiwyg/input.ts
@@ -2,14 +2,13 @@ import {isToC, renderToc} from "../util/fixBrowserBehavior";
 import {
     getTopList,
     hasClosestBlock, hasClosestByAttribute,
-    hasClosestByHeadings,
-    hasClosestByTag,
 } from "../util/hasClosest";
 import {log} from "../util/log";
 import {processCodeRender} from "../util/processCode";
 import {setRangeByWbr} from "../util/selection";
 import {afterRenderEvent} from "./afterRenderEvent";
 import {previoueIsEmptyA} from "./inlineTag";
+import {hasClosestByHeadings, hasClosestByTag} from "../util/hasClosestByHEadings";
 
 export const input = (vditor: IVditor, range: Range, event?: InputEvent) => {
     let blockElement = hasClosestBlock(range.startContainer);
@@ -39,7 +38,7 @@ export const input = (vditor: IVditor, range: Range, event?: InputEvent) => {
         }
 
         if (hasClosestByHeadings(blockElement)) {
-            renderToc(vditor.wysiwyg.element);
+            renderToc(vditor);
         }
 
         // 保存光标