+++ src/ts/wysiwyg/setHeading.ts
@@ -24,7 +24,7 @@ export const setHeading = (vditor: IVditor, tagName: string) => {
             blockElement.outerHTML = `<${tagName} data-block="0">${blockElement.innerHTML.trim()}</${tagName}>`;
         }
         setRangeByWbr(vditor.wysiwyg.element, range);
-        renderToc(vditor.wysiwyg.element);
+        renderToc(vditor);
     }
 };